Make `NodeId` non-comparable, reflecting that it's intended to be an opaque identifier for a node.
`PostorderIterator` and `SiblingIterator` still rely on the numerical
meaning of `NodeId`s, but that meaning is now an implementation detail
and no longer part of the API contract of the type.

Remove `operator<` from `SiblingIterator`; it's not expected for a
forward iterator to provide an ordering, and this operator was not used.