Switch to bazel_dep and normal zlib. (#3526)
The use of zlib-ng seemed to work previously, but may have only done so
because of some combination of an implementation detail of
`strip_include_prefix` in older versions of Bazel and some happenstance
of a system installed `zlib.h` being possible to find. With Bazel 7 it
started breaking on my machine which doesn't have a system `zlib.h`.

Using the "normal" zlib and the bzlmod rigging for it is simple and
seems to work.

Also adds the Linux AArch64 component of the Bazel lock file which was
necessary to test this on my Arm Linux machine.