Introduce custom hash table data structures.
The hash table design is heavily based on Abseil's ["Swiss
Tables"][swiss-tables] design. It uses an array of bytes storing
metadata about each entry and an array of entries where each is a pair
of key and value. The metadata byte consists of 7-bits of hash of the
key (distinct from the bits used to index the table), and one bit
indicating the presence of a special entry -- either empty or deleted.

[swiss-tables]: https://abseil.io/about/design/swisstables

See the comments in `raw_hashtable.h` for a detailed overview of the
design.

Co-authored-by: josh11b <15258583+josh11b@users.noreply.github.com>